+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - change to relative path outputs in log
 - suggest cd command to user when new ticket created
 - change parameter projectName to project\_name in .organizations json file
+- extractOrganizationFromUrl works with azure dev ops urls
 
 ### Removed